wildfire_clf/pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["wildfire_clf*"]

[project]
name = "wildfire-clf"
version = "0.1.0"
description = "Wildfire classification model serving API"
requires-python = ">=3.10"

dependencies = [
  # =========================
  # Core ML
  # =========================
  "torch==2.1.2",
  "torchvision==0.16.2",
  "timm==1.0.26",
  "numpy==1.26.4",
  "pillow==10.4.0",

  # =========================
  # API Serving
  # =========================
  "fastapi==0.135.3",
  "uvicorn[standard]==0.30.6",
  "python-multipart==0.0.9",
  "pydantic==2.7.4",
  "jinja2==3.1.4",

  # =========================
  # Utilities
  # =========================
  "tqdm==4.66.4",
  "pyyaml==6.0.1",
  "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1",
  "ruff>=0.4",
  "black>=24.4"
]

[project.scripts]
wildfire-serve = "wildfire_clf.serve:main"
wildfire-infer = "wildfire_clf.infer:main"